d['key'] instead of d.get('key') and refactor
BIG WARN: this changes are not tested

The first one is use of d['key'] instead of d.get('key').
The main difference is that d['key'] throw an error if not found. d.get('key',default_value) return default_value if not found. 
and d.get('key') is equivalent to d.get('key',None) so return None if not found.

The second one is a little refactor to remove violation to various DRY principle (line 82 to 87 of the new file)